14/02/24 - Resource Gathering complete
Player can now pick up resources and put them in its Handbag (child obj). An adjustable maximum is also included.

The Handbag is cleared when the player moves to a new scene. While this is an unintentional effect of unity, it fits quite well with the intentions for the gameplay.