21/04/24 Resources now dropped by Interactable
The Resource spawner now spawns a mix of 4 objects based on the old 'interactable'. These interactables take an object to spawn, a maximum amount to how many can spawn (1 - x) and the amount of times they need to be acted upon before they break.

The resources are then randomly spawned around the area relative to the position of the interactable instance.